Refactor water quality layer to use service
